Get and print a single character until 'q' is entered
The global variable `STDIN` is a handle on the standard input in Ruby.
It is an instance of the `IO` class. This class looks to handle much of
the work of getting the terminal into raw and out of raw mode, moving
the cursor etc. For now it looks like the `ruby-termios` dependency is
not required.