Enter and exit terminal raw mode
There are some default params that haven't been set for `raw!`. Because
of this only a single byte will be read when a read operation is
performed. This suits me for now.

`getch` reads and returns a character in raw mode. It has the same
params and defaults as `raw!`.

At the end I presume `cooked!` returns the terminal to normal mode give
cooked is the opposite of raw.

Strictly speaking manually setting to raw mode and going back to cooked
isn't required since `getch` also reads in raw mode. For now I'll leave
it this way since it is quite clear what mode we are in and there may
be other operations that are added which I want to be in raw mode for.