Add carriage returns manually to printing
Going into raw mode removes some output processing. The terminal was
translating any newline ('\n') as a carriage return followed by a
newline ('\r\n'). This had the effect of keeping the cursor moving to
the next column with each newline. This change returns the cursor to the
far left column on each newline.