Refactor raw mode entry
`raw` can take a block where it yields `self` in raw mode. This is a bit
more Ruby like and makes it clear that everything in the block is to be
done in raw mode.

I've left the `cooked!` send as a guarantee that the users terminal
returns to normal when the program ends.