Ensure raw mode is exited when program ends
Using `exit(0)` inside `process_keypress` means that the thread of
execution will never return to `main`. That make it possible that the
terminal will still be in raw mode when the program exits.

I prefer having `exit` at the end of `main` as that is where the
programs primary exit should be. This also cuts down on the side effects
in `process_keypress`.

Strictly speaking the returned character from `process_keypress` is not
needed right now but it does provide a convenient way to exit the loop
in `main` when `nil` is returned.