Use an append buffer to minimize the number of STDOUT writes
Before `draw_rows` was making a bunch of small writes every time the
screen is refreshed. This is a potential source of screen flicker if
there are unpredictable pauses between writes.

I could have used a normal Ruby string as the append buffer but instead
I made a buffer where the implementation is an array of single string
characters. In Ruby 3 strings are frozen by default, so appending to a
string is going to make a completely new string object. By using an
array of characters instead I am assuming that the performance impact of
joining the array of characters once when I am ready to write will be
better than appending to the same string variable.