After refreshing the screen return the cursor to the current position
A few things here:

We keep track of the editors cursor position with instance variables.

The IO library already seems to convert to the 1 based indexing that is
normal for a terminal screen coordinate. We only need to worry about 0
based indexing in the code.

The coordinates are listed back to front compared to normal convention:
the y coordinate comes first followed by the x. This is rows vs. columns
respectively.