Move cursor with arrow keys
This change detects the ASCII escape character for VT1000 terminals: 27.
Arrow keys have a sequence of bytes after an escape character e.g. the
left arrow is `[D`.

After the escape character is detected the next two bytes in the
sequence are checked to see which key was pressed.

I think in C a `char` and it's corresponding `int` can be considered
equal in that the character has the same underlying code representation.
This won't work in Ruby where even a single character is a String. So
in `read_key` I consistently handle each character by its ordinal value
i.e. the integer code representation. Using the string value aids
readability.