Implement page_up and page_down keys
For now it's hard to tell if these keys are working. Without a page of
text to scroll moving page is not obvious.

I've chosen 1 less row that the screen rows so the top or bottom row of
a screen will still be visible on the next page, to give the reader
continuity.