Display a line of text
The `open` method will open files in the text editor. For the moment it
is a hard coded string.

I've chosen for now to treat `@row` as an array of single characters
(each a string). This means that they have to be joined together before
adding them to the append buffer. I have treated it that way because I
may have been treating strings similarly elsewhere, I just can't
remember where.